Restore: refine split code for both compacted sst files and logs (#56917)
ref pingcap/tidb#56522
