+++ br/pkg/restore/snap_client/BUILD.bazel
@@ -26,7 +26,6 @@ go_library(
         "//br/pkg/restore/internal/import_client",
         "//br/pkg/restore/internal/prealloc_db",
         "//br/pkg/restore/internal/prealloc_table_id",
-        "//br/pkg/restore/internal/snap_split",
         "//br/pkg/restore/split",
         "//br/pkg/restore/utils",
         "//br/pkg/storage",
@@ -93,9 +92,9 @@ go_test(
         "//br/pkg/mock",
         "//br/pkg/restore",
         "//br/pkg/restore/internal/import_client",
+        "//br/pkg/restore/split",
         "//br/pkg/restore/utils",
         "//br/pkg/utils",
-        "//br/pkg/utiltest",
         "//pkg/domain",
         "//pkg/kv",
         "//pkg/meta/model",
