+++ br/pkg/restore/split/BUILD.bazel
@@ -7,6 +7,7 @@ go_library(
         "mock_pd_client.go",
         "region.go",
         "split.go",
+        "sum_sorted.go",
     ],
     importpath = "github.com/pingcap/tidb/br/pkg/restore/split",
     visibility = ["//visibility:public"],
@@ -14,19 +15,24 @@ go_library(
         "//br/pkg/conn/util",
         "//br/pkg/errors",
         "//br/pkg/logutil",
+        "//br/pkg/pdutil",
+        "//br/pkg/restore/utils",
         "//br/pkg/utils",
         "//pkg/kv",
         "//pkg/lightning/common",
         "//pkg/lightning/config",
         "//pkg/lightning/log",
         "//pkg/store/pdtypes",
+        "//pkg/tablecodec",
         "//pkg/util",
         "//pkg/util/codec",
         "//pkg/util/intest",
         "//pkg/util/redact",
         "@com_github_docker_go_units//:go-units",
+        "@com_github_google_btree//:btree",
         "@com_github_pingcap_errors//:errors",
         "@com_github_pingcap_failpoint//:failpoint",
+        "@com_github_pingcap_kvproto//pkg/brpb",
         "@com_github_pingcap_kvproto//pkg/errorpb",
         "@com_github_pingcap_kvproto//pkg/kvrpcpb",
         "@com_github_pingcap_kvproto//pkg/metapb",
@@ -39,6 +45,7 @@ go_library(
         "@org_golang_google_grpc//codes",
         "@org_golang_google_grpc//credentials",
         "@org_golang_google_grpc//credentials/insecure",
+        "@org_golang_google_grpc//keepalive",
         "@org_golang_google_grpc//status",
         "@org_golang_x_sync//errgroup",
         "@org_uber_go_multierr//:multierr",
@@ -52,12 +59,15 @@ go_test(
     srcs = [
         "client_test.go",
         "split_test.go",
+        "sum_sorted_test.go",
     ],
     embed = [":split"],
     flaky = True,
-    shard_count = 19,
+    shard_count = 27,
     deps = [
         "//br/pkg/errors",
+        "//br/pkg/restore/utils",
+        "//br/pkg/rtree",
         "//br/pkg/utils",
         "//pkg/kv",
         "//pkg/sessionctx/stmtctx",
@@ -65,8 +75,11 @@ go_test(
         "//pkg/tablecodec",
         "//pkg/types",
         "//pkg/util/codec",
+        "@com_github_docker_go_units//:go-units",
         "@com_github_pingcap_errors//:errors",
         "@com_github_pingcap_failpoint//:failpoint",
+        "@com_github_pingcap_kvproto//pkg/brpb",
+        "@com_github_pingcap_kvproto//pkg/import_sstpb",
         "@com_github_pingcap_kvproto//pkg/kvrpcpb",
         "@com_github_pingcap_kvproto//pkg/metapb",
         "@com_github_pingcap_kvproto//pkg/pdpb",
