+++ br/pkg/restore/split/split_test.go
@@ -5,16 +5,23 @@ import (
 	"bytes"
 	"context"
 	goerrors "errors"
+	"fmt"
 	"slices"
+	"sort"
 	"testing"
 	"time"
 
+	"github.com/docker/go-units"
 	"github.com/pingcap/errors"
 	"github.com/pingcap/failpoint"
+	backuppb "github.com/pingcap/kvproto/pkg/brpb"
+	"github.com/pingcap/kvproto/pkg/import_sstpb"
 	"github.com/pingcap/kvproto/pkg/kvrpcpb"
 	"github.com/pingcap/kvproto/pkg/metapb"
 	"github.com/pingcap/kvproto/pkg/pdpb"
 	berrors "github.com/pingcap/tidb/br/pkg/errors"
+	restoreutils "github.com/pingcap/tidb/br/pkg/restore/utils"
+	"github.com/pingcap/tidb/br/pkg/rtree"
 	"github.com/pingcap/tidb/br/pkg/utils"
 	"github.com/pingcap/tidb/pkg/kv"
 	"github.com/pingcap/tidb/pkg/store/pdtypes"
@@ -754,5 +761,353 @@ func TestScanRegionsWithRetry(t *testing.T) {
 		require.Len(t, regions, 2)
 		require.Equal(t, []byte("1"), regions[0].Region.StartKey)
 		require.Equal(t, []byte("2"), regions[1].Region.StartKey)
+	}
+}
+
+func TestScanEmptyRegion(t *testing.T) {
+	mockPDCli := NewMockPDClientForSplit()
+	mockPDCli.SetRegions([][]byte{{}, {12}, {34}, {}})
+	client := NewClient(mockPDCli, nil, nil, 100, 4)
+	keys := initKeys()
+	// make keys has only one
+	keys = keys[0:1]
+	regionSplitter := NewRegionSplitter(client)
+
+	ctx := context.Background()
+	err := regionSplitter.ExecuteSortedKeys(ctx, keys)
+	// should not return error with only one range entry
+	require.NoError(t, err)
+}
+
+func TestSplitEmptyRegion(t *testing.T) {
+	mockPDCli := NewMockPDClientForSplit()
+	mockPDCli.SetRegions([][]byte{{}, {12}, {34}, {}})
+	client := NewClient(mockPDCli, nil, nil, 100, 4)
+	regionSplitter := NewRegionSplitter(client)
+	err := regionSplitter.ExecuteSortedKeys(context.Background(), nil)
+	require.NoError(t, err)
+}
+
+// region: [, aay), [aay, bba), [bba, bbh), [bbh, cca), [cca, )
+// range: [aaa, aae), [aae, aaz), [ccd, ccf), [ccf, ccj)
+// rewrite rules: aa -> xx,  cc -> bb
+// expected regions after split:
+//
+//	[, aay), [aay, bba), [bba, bbf), [bbf, bbh), [bbh, bbj),
+//	[bbj, cca), [cca, xxe), [xxe, xxz), [xxz, )
+func TestSplitAndScatter(t *testing.T) {
+	rangeBoundaries := [][]byte{[]byte(""), []byte("aay"), []byte("bba"), []byte("bbh"), []byte("cca"), []byte("")}
+	encodeBytes(rangeBoundaries)
+	mockPDCli := NewMockPDClientForSplit()
+	mockPDCli.SetRegions(rangeBoundaries)
+	client := NewClient(mockPDCli, nil, nil, 100, 4)
+	regionSplitter := NewRegionSplitter(client)
+	ctx := context.Background()
+
+	ranges := initRanges()
+	rules := initRewriteRules()
+	splitKeys := make([][]byte, 0, len(ranges))
+	for _, rg := range ranges {
+		tmp, err := restoreutils.RewriteRange(&rg, rules)
+		require.NoError(t, err)
+		splitKeys = append(splitKeys, tmp.EndKey)
+	}
+	sort.Slice(splitKeys, func(i, j int) bool {
+		return bytes.Compare(splitKeys[i], splitKeys[j]) < 0
+	})
+	err := regionSplitter.ExecuteSortedKeys(ctx, splitKeys)
+	require.NoError(t, err)
+	regions := mockPDCli.Regions.ScanRange(nil, nil, 100)
+	expected := [][]byte{[]byte(""), []byte("aay"), []byte("bba"), []byte("bbf"), []byte("bbh"), []byte("bbj"), []byte("cca"), []byte("xxe"), []byte("xxz"), []byte("")}
+	encodeBytes(expected)
+	require.Len(t, regions, len(expected)-1)
+	for i, region := range regions {
+		require.Equal(t, expected[i], region.Meta.StartKey)
+		require.Equal(t, expected[i+1], region.Meta.EndKey)
+	}
+}
+
+func encodeBytes(keys [][]byte) {
+	for i := range keys {
+		if len(keys[i]) == 0 {
+			continue
+		}
+		keys[i] = codec.EncodeBytes(nil, keys[i])
+	}
+}
+
+func TestRawSplit(t *testing.T) {
+	// Fix issue #36490.
+	splitKeys := [][]byte{{}}
+	ctx := context.Background()
+	rangeBoundaries := [][]byte{[]byte(""), []byte("aay"), []byte("bba"), []byte("bbh"), []byte("cca"), []byte("")}
+	mockPDCli := NewMockPDClientForSplit()
+	mockPDCli.SetRegions(rangeBoundaries)
+	client := NewClient(mockPDCli, nil, nil, 100, 4, WithRawKV())
+
+	regionSplitter := NewRegionSplitter(client)
+	err := regionSplitter.ExecuteSortedKeys(ctx, splitKeys)
+	require.NoError(t, err)
+
+	regions := mockPDCli.Regions.ScanRange(nil, nil, 100)
+	require.Len(t, regions, len(rangeBoundaries)-1)
+	for i, region := range regions {
+		require.Equal(t, rangeBoundaries[i], region.Meta.StartKey)
+		require.Equal(t, rangeBoundaries[i+1], region.Meta.EndKey)
+	}
+}
+
+// keys: aae, aaz, ccf, ccj
+func initKeys() [][]byte {
+	return [][]byte{
+		[]byte("aae"),
+		[]byte("aaz"),
+		[]byte("ccf"),
+		[]byte("ccj"),
+	}
+}
+
+// range: [aaa, aae), [aae, aaz), [ccd, ccf), [ccf, ccj)
+func initRanges() []rtree.Range {
+	var ranges [4]rtree.Range
+	ranges[0] = rtree.Range{
+		StartKey: []byte("aaa"),
+		EndKey:   []byte("aae"),
+	}
+	ranges[1] = rtree.Range{
+		StartKey: []byte("aae"),
+		EndKey:   []byte("aaz"),
+	}
+	ranges[2] = rtree.Range{
+		StartKey: []byte("ccd"),
+		EndKey:   []byte("ccf"),
+	}
+	ranges[3] = rtree.Range{
+		StartKey: []byte("ccf"),
+		EndKey:   []byte("ccj"),
+	}
+	return ranges[:]
+}
+
+func initRewriteRules() *restoreutils.RewriteRules {
+	var rules [2]*import_sstpb.RewriteRule
+	rules[0] = &import_sstpb.RewriteRule{
+		OldKeyPrefix: []byte("aa"),
+		NewKeyPrefix: []byte("xx"),
+	}
+	rules[1] = &import_sstpb.RewriteRule{
+		OldKeyPrefix: []byte("cc"),
+		NewKeyPrefix: []byte("bb"),
+	}
+	return &restoreutils.RewriteRules{
+		Data: rules[:],
+	}
+}
+
+func keyWithTablePrefix(tableID int64, key string) []byte {
+	rawKey := append(tablecodec.GenTableRecordPrefix(tableID), []byte(key)...)
+	return codec.EncodeBytes([]byte{}, rawKey)
+}
+
+func TestSplitPoint(t *testing.T) {
+	ctx := context.Background()
+	var oldTableID int64 = 50
+	var tableID int64 = 100
+	rewriteRules := &restoreutils.RewriteRules{
+		Data: []*import_sstpb.RewriteRule{
+			{
+				OldKeyPrefix: tablecodec.EncodeTablePrefix(oldTableID),
+				NewKeyPrefix: tablecodec.EncodeTablePrefix(tableID),
+			},
+		},
+	}
+
+	// range:     b   c d   e       g         i
+	//            +---+ +---+       +---------+
+	//          +-------------+----------+---------+
+	// region:  a             f          h         j
+	splitHelper := NewSplitHelper()
+	splitHelper.Merge(Valued{Key: Span{StartKey: keyWithTablePrefix(oldTableID, "b"), EndKey: keyWithTablePrefix(oldTableID, "c")}, Value: Value{Size: 100, Number: 100}})
+	splitHelper.Merge(Valued{Key: Span{StartKey: keyWithTablePrefix(oldTableID, "d"), EndKey: keyWithTablePrefix(oldTableID, "e")}, Value: Value{Size: 200, Number: 200}})
+	splitHelper.Merge(Valued{Key: Span{StartKey: keyWithTablePrefix(oldTableID, "g"), EndKey: keyWithTablePrefix(oldTableID, "i")}, Value: Value{Size: 300, Number: 300}})
+	client := NewFakeSplitClient()
+	client.AppendRegion(keyWithTablePrefix(tableID, "a"), keyWithTablePrefix(tableID, "f"))
+	client.AppendRegion(keyWithTablePrefix(tableID, "f"), keyWithTablePrefix(tableID, "h"))
+	client.AppendRegion(keyWithTablePrefix(tableID, "h"), keyWithTablePrefix(tableID, "j"))
+	client.AppendRegion(keyWithTablePrefix(tableID, "j"), keyWithTablePrefix(tableID+1, "a"))
+
+	iter := NewSplitHelperIteratorForTest(splitHelper, tableID, rewriteRules)
+	err := SplitPoint(ctx, iter, client, func(ctx context.Context, rs *RegionSplitter, u uint64, o int64, ri *RegionInfo, v []Valued) error {
+		require.Equal(t, u, uint64(0))
+		require.Equal(t, o, int64(0))
+		require.Equal(t, ri.Region.StartKey, keyWithTablePrefix(tableID, "a"))
+		require.Equal(t, ri.Region.EndKey, keyWithTablePrefix(tableID, "f"))
+		require.EqualValues(t, v[0].Key.StartKey, keyWithTablePrefix(tableID, "b"))
+		require.EqualValues(t, v[0].Key.EndKey, keyWithTablePrefix(tableID, "c"))
+		require.EqualValues(t, v[1].Key.StartKey, keyWithTablePrefix(tableID, "d"))
+		require.EqualValues(t, v[1].Key.EndKey, keyWithTablePrefix(tableID, "e"))
+		require.Equal(t, len(v), 2)
+		return nil
+	})
+	require.NoError(t, err)
+}
+
+func getCharFromNumber(prefix string, i int) string {
+	c := '1' + (i % 10)
+	b := '1' + (i%100)/10
+	a := '1' + i/100
+	return fmt.Sprintf("%s%c%c%c", prefix, a, b, c)
+}
+
+func TestSplitPoint2(t *testing.T) {
+	ctx := context.Background()
+	var oldTableID int64 = 50
+	var tableID int64 = 100
+	rewriteRules := &restoreutils.RewriteRules{
+		Data: []*import_sstpb.RewriteRule{
+			{
+				OldKeyPrefix: tablecodec.EncodeTablePrefix(oldTableID),
+				NewKeyPrefix: tablecodec.EncodeTablePrefix(tableID),
+			},
+		},
+	}
+
+	// range:     b   c d   e f                 i j    k l        n
+	//            +---+ +---+ +-----------------+ +----+ +--------+
+	//          +---------------+--+.....+----+------------+---------+
+	// region:  a               g   >128      h            m         o
+	splitHelper := NewSplitHelper()
+	splitHelper.Merge(Valued{Key: Span{StartKey: keyWithTablePrefix(oldTableID, "b"), EndKey: keyWithTablePrefix(oldTableID, "c")}, Value: Value{Size: 100, Number: 100}})
+	splitHelper.Merge(Valued{Key: Span{StartKey: keyWithTablePrefix(oldTableID, "d"), EndKey: keyWithTablePrefix(oldTableID, "e")}, Value: Value{Size: 200, Number: 200}})
+	splitHelper.Merge(Valued{Key: Span{StartKey: keyWithTablePrefix(oldTableID, "f"), EndKey: keyWithTablePrefix(oldTableID, "i")}, Value: Value{Size: 300, Number: 300}})
+	splitHelper.Merge(Valued{Key: Span{StartKey: keyWithTablePrefix(oldTableID, "j"), EndKey: keyWithTablePrefix(oldTableID, "k")}, Value: Value{Size: 200, Number: 200}})
+	splitHelper.Merge(Valued{Key: Span{StartKey: keyWithTablePrefix(oldTableID, "l"), EndKey: keyWithTablePrefix(oldTableID, "n")}, Value: Value{Size: 200, Number: 200}})
+	client := NewFakeSplitClient()
+	client.AppendRegion(keyWithTablePrefix(tableID, "a"), keyWithTablePrefix(tableID, "g"))
+	client.AppendRegion(keyWithTablePrefix(tableID, "g"), keyWithTablePrefix(tableID, getCharFromNumber("g", 0)))
+	for i := 0; i < 256; i++ {
+		client.AppendRegion(keyWithTablePrefix(tableID, getCharFromNumber("g", i)), keyWithTablePrefix(tableID, getCharFromNumber("g", i+1)))
+	}
+	client.AppendRegion(keyWithTablePrefix(tableID, getCharFromNumber("g", 256)), keyWithTablePrefix(tableID, "h"))
+	client.AppendRegion(keyWithTablePrefix(tableID, "h"), keyWithTablePrefix(tableID, "m"))
+	client.AppendRegion(keyWithTablePrefix(tableID, "m"), keyWithTablePrefix(tableID, "o"))
+	client.AppendRegion(keyWithTablePrefix(tableID, "o"), keyWithTablePrefix(tableID+1, "a"))
+
+	firstSplit := true
+	iter := NewSplitHelperIteratorForTest(splitHelper, tableID, rewriteRules)
+	err := SplitPoint(ctx, iter, client, func(ctx context.Context, rs *RegionSplitter, u uint64, o int64, ri *RegionInfo, v []Valued) error {
+		if firstSplit {
+			require.Equal(t, u, uint64(0))
+			require.Equal(t, o, int64(0))
+			require.Equal(t, ri.Region.StartKey, keyWithTablePrefix(tableID, "a"))
+			require.Equal(t, ri.Region.EndKey, keyWithTablePrefix(tableID, "g"))
+			require.EqualValues(t, v[0].Key.StartKey, keyWithTablePrefix(tableID, "b"))
+			require.EqualValues(t, v[0].Key.EndKey, keyWithTablePrefix(tableID, "c"))
+			require.EqualValues(t, v[1].Key.StartKey, keyWithTablePrefix(tableID, "d"))
+			require.EqualValues(t, v[1].Key.EndKey, keyWithTablePrefix(tableID, "e"))
+			require.EqualValues(t, v[2].Key.StartKey, keyWithTablePrefix(tableID, "f"))
+			require.EqualValues(t, v[2].Key.EndKey, keyWithTablePrefix(tableID, "g"))
+			require.Equal(t, v[2].Value.Size, uint64(1))
+			require.Equal(t, v[2].Value.Number, int64(1))
+			require.Equal(t, len(v), 3)
+			firstSplit = false
+		} else {
+			require.Equal(t, u, uint64(1))
+			require.Equal(t, o, int64(1))
+			require.Equal(t, ri.Region.StartKey, keyWithTablePrefix(tableID, "h"))
+			require.Equal(t, ri.Region.EndKey, keyWithTablePrefix(tableID, "m"))
+			require.EqualValues(t, v[0].Key.StartKey, keyWithTablePrefix(tableID, "j"))
+			require.EqualValues(t, v[0].Key.EndKey, keyWithTablePrefix(tableID, "k"))
+			require.EqualValues(t, v[1].Key.StartKey, keyWithTablePrefix(tableID, "l"))
+			require.EqualValues(t, v[1].Key.EndKey, keyWithTablePrefix(tableID, "m"))
+			require.Equal(t, v[1].Value.Size, uint64(100))
+			require.Equal(t, v[1].Value.Number, int64(100))
+			require.Equal(t, len(v), 2)
+		}
+		return nil
+	})
+	require.NoError(t, err)
+}
+
+func fakeFile(tableID, rowID int64, length uint64, num int64) *backuppb.DataFileInfo {
+	return &backuppb.DataFileInfo{
+		StartKey:        fakeRowKey(tableID, rowID),
+		EndKey:          fakeRowKey(tableID, rowID+1),
+		TableId:         tableID,
+		Length:          length,
+		NumberOfEntries: num,
+	}
+}
+
+func fakeRowKey(tableID, rowID int64) kv.Key {
+	return codec.EncodeBytes(nil, tablecodec.EncodeRecordKey(tablecodec.GenTableRecordPrefix(tableID), kv.IntHandle(rowID)))
+}
+
+func TestLogSplitHelper(t *testing.T) {
+	ctx := context.Background()
+	rules := map[int64]*restoreutils.RewriteRules{
+		1: {
+			Data: []*import_sstpb.RewriteRule{
+				{
+					OldKeyPrefix: tablecodec.GenTableRecordPrefix(1),
+					NewKeyPrefix: tablecodec.GenTableRecordPrefix(100),
+				},
+			},
+		},
+		2: {
+			Data: []*import_sstpb.RewriteRule{
+				{
+					OldKeyPrefix: tablecodec.GenTableRecordPrefix(2),
+					NewKeyPrefix: tablecodec.GenTableRecordPrefix(200),
+				},
+			},
+		},
+	}
+	oriRegions := [][]byte{
+		{},
+		codec.EncodeBytes(nil, tablecodec.EncodeTablePrefix(100)),
+		codec.EncodeBytes(nil, tablecodec.EncodeTablePrefix(200)),
+		codec.EncodeBytes(nil, tablecodec.EncodeTablePrefix(402)),
+	}
+	mockPDCli := NewMockPDClientForSplit()
+	mockPDCli.SetRegions(oriRegions)
+	client := NewClient(mockPDCli, nil, nil, 100, 4)
+	helper := NewLogSplitHelper(rules, client, 4*units.MiB, 400)
+
+	helper.Merge(fakeFile(1, 100, 100, 100))
+	helper.Merge(fakeFile(1, 200, 2*units.MiB, 200))
+	helper.Merge(fakeFile(2, 100, 3*units.MiB, 300))
+	helper.Merge(fakeFile(3, 100, 10*units.MiB, 100000))
+	// different regions, no split happens
+	err := helper.Split(ctx)
+	require.NoError(t, err)
+	regions, err := mockPDCli.ScanRegions(ctx, []byte{}, []byte{}, 0)
+	require.NoError(t, err)
+	require.Len(t, regions, 3)
+	require.Equal(t, []byte{}, regions[0].Meta.StartKey)
+	require.Equal(t, codec.EncodeBytes(nil, tablecodec.EncodeTablePrefix(100)), regions[1].Meta.StartKey)
+	require.Equal(t, codec.EncodeBytes(nil, tablecodec.EncodeTablePrefix(200)), regions[2].Meta.StartKey)
+	require.Equal(t, codec.EncodeBytes(nil, tablecodec.EncodeTablePrefix(402)), regions[2].Meta.EndKey)
+
+	helper.Merge(fakeFile(1, 300, 3*units.MiB, 10))
+	helper.Merge(fakeFile(1, 400, 4*units.MiB, 10))
+	// trigger to split regions for table 1
+	err = helper.Split(ctx)
+	require.NoError(t, err)
+	regions, err = mockPDCli.ScanRegions(ctx, []byte{}, []byte{}, 0)
+	require.NoError(t, err)
+	require.Len(t, regions, 4)
+	require.Equal(t, fakeRowKey(100, 400), kv.Key(regions[1].Meta.EndKey))
+}
+
+func NewSplitHelperIteratorForTest(helper *SplitHelper, tableID int64, rule *restoreutils.RewriteRules) *splitHelperIterator {
+	return &splitHelperIterator{
+		tableSplitters: []*rewriteSplitter{
+			{
+				tableID:  tableID,
+				rule:     rule,
+				splitter: helper,
+			},
+		},
 	}
 }
