+++ br/pkg/restore/split/sum_sorted.go
@@ -1,5 +1,5 @@
 // Copyright 2022 PingCAP, Inc. Licensed under Apache-2.0.
-package logsplit
+package split
 
 import (
 	"bytes"
