+++ br/pkg/utiltest/BUILD.bazel
@@ -2,23 +2,13 @@ load("@io_bazel_rules_go//go:def.bzl", "go_library")
 
 go_library(
     name = "utiltest",
-    srcs = [
-        "fake.go",
-        "suite.go",
-    ],
+    srcs = ["suite.go"],
     importpath = "github.com/pingcap/tidb/br/pkg/utiltest",
     visibility = ["//visibility:public"],
     deps = [
         "//br/pkg/gluetidb/mock",
         "//br/pkg/mock",
-        "//br/pkg/pdutil",
-        "//br/pkg/restore/split",
         "//br/pkg/storage",
-        "@com_github_pingcap_kvproto//pkg/metapb",
-        "@com_github_pkg_errors//:errors",
         "@com_github_stretchr_testify//require",
-        "@com_github_tikv_pd_client//:client",
-        "@com_github_tikv_pd_client//http",
-        "@org_golang_google_grpc//keepalive",
     ],
 )
